refactor: [review] Response(controller), Request(domain) 분리
